Move KeyValueStore to key_value_store.py.
